# swagger.yaml
openapi: 3.0.0
info:
  title: 'API Livros'
  description: 'Documentação da API Livros'
  version: 1.0.0
servers:
  -
    url: 'http://localhost/Api-Livros'
    description: 'Servidor local'
paths:
...
  /livros:
    put:
      tags:
        - Livros
      summary: 'Atualiza um livro existente do usuário autenticado'
      operationId: ee928fab664b23084de1dbe7464ea81e
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - id_livro
              properties:
                id_livro:
                  type: integer
                  example: 10
                titulo:
                  type: string
                  example: 'Dom Casmurro - edição revisada'
                  nullable: true
                autor:
                  type: string
                  example: 'Machado de Assis'
                  nullable: true
                ano:
                  type: integer
                  example: 1900
                  nullable: true
                genero:
                  type: string
                  example: Romance
                  nullable: true
                status:
                  type: string
                  example: lendo
                  nullable: true
                avaliacao:
                  type: integer
                  example: 4
                  nullable: true
                anotacoes:
                  type: string
                  example: 'Leitura em andamento'
                  nullable: true
              type: object
      responses:
        '200':
          description: 'Livro atualizado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Livro atualizado com sucesso' }
                  detail: { properties: { livro: { properties: { id: { type: integer, example: 10 }, titulo: { type: string, example: 'Dom Casmurro - edição revisada' }, autor: { type: string, example: 'Machado de Assis' }, ano: { type: integer, example: 1900 }, genero: { type: string, example: Romance, nullable: true }, status: { type: string, example: lendo }, avaliacao: { type: integer, example: 4, nullable: true }, anotacoes: { type: string, example: 'Leitura em andamento', nullable: true } }, type: object } }, type: object }
                type: object
        '400':
          description: 'Id do livro ausente ou dados inválidos'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Id do livro é obrigatório' }
                type: object
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Livro não encontrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Livro não encontrado' }
                type: object
        '500':
          description: 'Erro ao atualizar livro'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao atualizar livro' }
                type: object
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Livros
      summary: 'Cria um novo livro para o usuário autenticado'
      operationId: b739368170d28460413bda43250d382b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - titulo
                - autor
                - ano
                - status
              properties:
                titulo:
                  type: string
                  example: 'Dom Casmurro'
                autor:
                  type: string
                  example: 'Machado de Assis'
                ano:
                  type: integer
                  example: 1899
                genero:
                  type: string
                  example: Romance
                  nullable: true
                status:
                  type: string
                  example: quero_ler
                avaliacao:
                  type: integer
                  example: 5
                  nullable: true
                anotacoes:
                  type: string
                  example: 'Clássico da literatura brasileira'
                  nullable: true
              type: object
      responses:
        '201':
          description: 'Livro criado com sucesso'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Livro criado com sucesso' }
                  id: { type: integer, example: 10 }
                type: object
        '400':
          description: 'Dados inválidos'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Campo título é obrigatório' }
                type: object
        '401':
          description: 'Não autenticado'
        '500':
          description: 'Erro ao criar livro'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Erro ao criar livro' }
                type: object
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Livros
      summary: 'Deleta um livro do usuário autenticado'
      operationId: d4487ace35a2539f559bc45a6a2a4568
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - id_livro
              properties:
                id_livro:
                  type: integer
                  example: 10
              type: object
      responses:
        '204':
          description: 'Livro deletado com sucesso'
        '400':
          description: 'Id do livro não informado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Id do livro é obrigatório' }
                type: object
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Livro não encontrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Livro não encontrado' }
                type: object
      security:
        -
          bearerAuth: []
  /livros/listar:
    get:
      tags:
        - Livros
      summary: 'Lista livros do usuário autenticado'
      operationId: 7e1bfe193830cf536c5ae807c7be522e
      parameters:
        -
          name: page
          in: query
          description: 'Página atual'
          required: false
          schema:
            type: integer
            example: 1
            default: 1
        -
          name: limit
          in: query
          description: 'Quantidade de itens por página'
          required: false
          schema:
            type: integer
            example: 10
            default: 10
      responses:
        '200':
          description: 'Livros retornados com sucesso'
          content:
            application/json:
              schema:
                properties:
                  detail: { properties: { livros: { type: array, items: { properties: { titulo: { type: string, example: 'Dom Casmurro' }, autor: { type: string, example: 'Machado de Assis' }, ano: { type: integer, example: 1899, nullable: true }, genero: { type: string, example: Romance }, status: { type: string, example: lendo }, avaliacao: { type: integer, example: 5, nullable: true }, anotacoes: { type: string, example: 'Muito bom', nullable: true } }, type: object } }, paginacao: { properties: { page: { type: integer, example: 1 }, limit: { type: integer, example: 10 }, total: { type: integer, example: 1 }, total_pages: { type: integer, example: 1 } }, type: object } }, type: object }
                type: object
        '401':
          description: 'Não autenticado'
        '404':
          description: 'Nenhum livro encontrado'
          content:
            application/json:
              schema:
                properties:
                  mensagem: { type: string, example: 'Nenhum livro encontrado' }
                type: object
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...